refactor(daemon): extract Task data models from runner.py (#798)
Extract Task, TaskStatus, TaskKind, DuplicateTaskIdError,
QueueSaturationError, and DaemonState into a dedicated
maxwell_daemon/daemon/task_models.py module. All names remain
importable from runner.py via re-export for backwards compatibility.

runner.py shrinks from 2,141 → 2,056 lines.

Partially addresses #798
